+++ src/index.ts
@@ -1,4 +1,6 @@
 import createAsyncAction from './createAsyncAction';
 import reducer from './reducer';
+import AwaitControl from './AwaitControl';
 
 export { createAsyncAction, reducer };
+export default AwaitControl;