+++ src/types.ts
@@ -1,8 +1,10 @@
-import { Action, ActionFunctionAny } from 'redux-actions';
+import { Action, ActionFunctionAny, ActionMeta } from 'redux-actions';
 
 export type AsyncActionState = 'START' | 'SUCCESS' | 'FAILURE' | 'CANCEL';
 
-export type AsyncBaseAction = {
+export type BaseAction = Action<any> | ActionMeta<any, any>;
+
+export type AsyncBaseActionControl = {
   start: ActionFunctionAny<Action<any>>;
   success: ActionFunctionAny<Action<any>>;
   failure: ActionFunctionAny<Action<any>>;
@@ -11,4 +13,10 @@ export type AsyncBaseAction = {
 
 export type AsyncActionReducer = {
   [actionName: string]: AsyncActionState;
+};
+
+export type Selector = (state: any) => any;
+
+export type AwaitControlInitializer = {
+  keyReducer?: string;
 };