+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Minor Changes
 
-- Now it`s possible to add a key for action, then it`s possible controlling some actions of the same type
+- Now it's possible to add a key for action, then it's possible controlling some actions of the same type
 
 ## 0.2.0